requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
timm==0.9.12

# GBDT Libraries
lightgbm==4.1.0
xgboost==2.0.3
catboost==1.2.2

# Hyperparameter Optimization
optuna==3.5.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
Pillow==10.1.0

# Feature Engineering
regex==2023.10.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Progress and Logging
tqdm==4.66.1

# Testing
pytest==7.4.3
hypothesis==6.92.1

# Utilities
pyyaml==6.0.1
requests==2.31.0

# Parameter Efficient Fine-tuning
peft==0.7.1
